Bug-For numbers biger than 3, MilerRabin returns false

+++ IzbraniAlgVaja1/PrimeRandomNumberGenerator.h
@@ -7,6 +7,6 @@ public:
 	~PrimeRandomNumberGenerator();
 	long* Naive();
 	bool isPrime(long long num);
-	bool MilerRabin(long p,long s);
+	bool MilerRabin(long long  p,long s);
 };
